requirements.txt
langchain
openai
langchain-community
python-dotenv
streamlit
fastapi
langserve
uvicorn
sse_starlette
